moduling the code - first steps
